requirements.txt
# 自媒体博主自动化辅助平台依赖包

# GUI框架
PyQt5==5.15.9

# 后端框架
fastapi==0.104.1
uvicorn==0.23.2

# 数据库
sqlalchemy==2.0.23
sqlite3==0.0.1

# 爬虫
scrapy==2.11.0
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0

# 自然语言处理
langchain==0.0.335
langchain-community==0.0.10
sentence-transformers==2.2.2

# LLM API客户端
openai==1.3.5

# 工具库
numpy==1.26.2
pandas==2.1.3
pydantic==2.5.2
python-dotenv==1.0.0
tqdm==4.66.1 